Add files via upload
Transportation is a major issue to both students and parents. Most of the university students are matured enough to travel their campuses by themselves. In these cases, students use public transport from different places. DIU is always concerned about their students and for this purpose now, We have around 50 + buses which are exclusively used for our students/faculty/admin personnel and staff. A parents Car and an ambulance is also available for the students and their parents.  It is time consuming and also saves money. Particulerly this service is very helpful for our female students. They can easily avoid the unavoidable circumstances or any types of unpleasant situations.